Working on Phase 1
Finished thrusters and shield strength portions of phase 1.  Currently
working on health collectible.  Already created a health pickup for
this, but need to find out how to turn it into a sprite that animates.
From there, add code that heals the player by one and ignores the player
if health is already full